# api/templates/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MetaTrader 5 Flask REST API",
    "description": "Comprehensive REST API exposing 100% of the official Python MetaTrader 5 library methods in a headless Wine container.\n\n### Modules\n- **System**: MT5 lifecycle, connection, login, terminal and account status.\n- **Symbols & DOM**: Symbol listings, specifications, real-time ticks, and Depth of Market.\n- **Market Data**: Historical rates (OHLCV) and ticks with flexible filtering.\n- **Trading**: Market/pending order execution, positions management, margin & profit calculators.\n- **History**: Closed orders and deals history with date and ticket filters.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local API Server"
    }
  ],
...
  "paths": {
...
    "/api/copy_rates_from": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Copy Rates From Date",
        "description": "Returns historical bars starting from a specified date backward/forward.",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EURUSD"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "M1",
              "enum": [
                "M1",
                "M2",
                "M3",
                "M4",
                "M5",
                "M6",
                "M10",
                "M12",
                "M15",
                "M20",
                "M30",
                "H1",
                "H2",
                "H3",
                "H4",
                "H6",
                "H8",
                "H12",
                "D1",
                "W1",
                "MN1"
              ]
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-01T00:00:00Z"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candlestick rates list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/copy_rates_from_pos": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Copy Rates From Position (Index)",
        "description": "Returns historical bars starting from an index offset (`start_pos=0` is current open candle).",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EURUSD"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "M1"
            }
          },
          {
            "name": "start_pos",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candlestick rates list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/copy_rates_range": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Copy Rates in Date Range",
        "description": "Returns historical bars within a specific start and end date range.",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EURUSD"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "M1"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-01T00:00:00Z"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-02T00:00:00Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candlestick rates list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/copy_ticks_from": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Copy Ticks From Date",
        "description": "Returns tick history starting from a specified date.",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EURUSD"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-01T00:00:00Z"
            }
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "flags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ALL",
              "enum": [
                "ALL",
                "INFO",
                "TRADE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticks list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/copy_ticks_range": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "summary": "Copy Ticks in Date Range",
        "description": "Returns tick history within a start and end datetime range.",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EURUSD"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-01T00:00:00Z"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-02T00:00:00Z"
            }
          },
          {
            "name": "flags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ALL",
              "enum": [
                "ALL",
                "INFO",
                "TRADE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticks list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/history_orders_total": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get Total Historical Orders Count",
        "description": "Returns count of closed/cancelled orders within the given time range.",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2026-01-01T00:00:00Z"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2026-01-31T23:59:59Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical orders count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history_orders_get": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get Historical Orders",
        "description": "Returns list of historical orders filtered by date range, group, order ticket, or position ticket.",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "position",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical orders list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history_deals_total": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get Total Historical Deals Count",
        "description": "Returns count of trade execution deals in history within the time window.",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical deals count",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/history_deals_get": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get Historical Deals (Fills & Closed Trades)",
        "description": "Returns list of executed deals with entry/exit prices, commission, swap, and realized profit.",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "position",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical deals list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    }
  }
}